# openapi/spec.yml
openapi: 3.0.0
info:
  contact:
    name: Canva Developer Community
    url: https://community.canva.dev/
  description: API for building integrations with Canva via a REST api
  license:
    name: ©2023 All Rights Reserved
  termsOfService: https://www.canva.com/trust/legal/
  title: Canva Connect API
  version: latest
servers:
  - description: Canva Connect API
    url: https://api.canva.com/rest
...
paths:
...
  /v1/oauth/token:
    post:
      description: |-
        This endpoint implements the OAuth 2.0 `token` endpoint, as part of the Authorization Code flow with Proof Key for Code Exchange (PKCE). For more information, see [Authentication](https://www.canva.dev/docs/connect/authentication/).

        To generate an access token, you must provide one of the following:

        - An authorization code
        - A refresh token

        Generating a token using either an authorization code or a refresh token allows your integration to act on behalf of a user. You must first [obtain user authorization and get an authorization code](https://www.canva.dev/docs/connect/authentication/#obtain-user-authorization).

        Access tokens may be up to 4 KB in size, and are only valid for a specified period of time. The expiry time (currently 4 hours) is shown in the endpoint response and is subject to change.

        **Endpoint authentication**

        Requests to this endpoint require authentication with your client ID and client secret, using _one_ of the following methods:

        - **Basic access authentication** (Recommended): For [basic access authentication](https://en.wikipedia.org/wiki/Basic_access_authentication), the `{credentials}` string must be a Base64 encoded value of `{client id}:{client secret}`.
        - **Body parameters**: Provide your integration's credentials using the `client_id` and `client_secret` body parameters.

        **Generate an access token using an authorization code**

        To generate an access token with an authorization code, you must:

        - Set `grant_type` to `authorization_code`.
        - Provide the `code_verifier` value that you generated when creating the user authorization URL.
        - Provide the authorization code you received after the user authorized the integration.

        **Generate an access token using a refresh token**

        Using the `refresh_token` value from a previous user token request, you can get a new access token with the same or smaller scope as the previous one, but with a refreshed expiry time. You will also receive a new refresh token that you can use to refresh the access token again.

        To refresh an existing access token, you must:

        - Set `grant_type` to `refresh_token`.
        - Provide the `refresh_token` from a previous token request.
      operationId: exchangeAccessToken
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ExchangeAccessTokenRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeAccessTokenResponse'
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OauthError'
          description: Error Response
      security:
        - {}
        - basicAuth: []
      tags:
        - oauth
...
  /v1/oauth/revoke:
    post:
      description: |-
        Revoke an access token or a refresh token.

        If you revoke a _refresh token_, be aware that:

        - The refresh token's lineage is also revoked. This means that access tokens created from that refresh token are also revoked.
        - The user's consent for your integration is also revoked. This means that the user must go through the OAuth process again to use your integration.

        Requests to this endpoint require authentication with your client ID and client secret, using _one_ of the following methods:

        - **Basic access authentication** (Recommended): For [basic access authentication](https://en.wikipedia.org/wiki/Basic_access_authentication), the `{credentials}` string must be a Base64 encoded value of `{client id}:{client secret}`.
        - **Body parameters**: Provide your integration's credentials using the `client_id` and `client_secret` body parameters.
      operationId: revokeTokens
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RevokeTokensRequest'
        required: true
      responses:
        "204":
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OauthError'
          description: Error Response
      security:
        - {}
        - basicAuth: []
      tags:
        - oauth
...
components:
...
  schemas:
...
    OauthError:
      properties:
        error:
          $ref: '#/components/schemas/ErrorCode'
        error_description:
          description: A human-readable description of what went wrong.
          type: string
      required:
        - error
        - error_description
      type: object
    ErrorCode:
      description: |
        A short string indicating what failed. This field can be used to handle errors programmatically.
      enum:
        - internal_error
        - invalid_field
        - invalid_header_value
        - permission_denied
        - too_many_requests
        - not_found
        - bad_request_body
        - bad_http_method
        - bad_request_params
        - bad_query_params
        - endpoint_not_found
        - unsupported_version
        - invalid_access_token
        - revoked_access_token
        - missing_field
        - missing_scope
        - invalid_grant
        - invalid_request
        - invalid_client
        - unauthorized_client
        - unsupported_grant_type
        - invalid_scope
        - invalid_basic_header
        - invalid_file_format
        - quota_exceeded
        - unsupported_content_type
        - request_too_large
        - folder_not_found
        - asset_not_found
        - max_limit_reached
        - permission_not_found
        - permission_exists
        - unauthorized_user
        - user_not_found
        - group_not_found
        - app_not_found
        - content_not_found
        - doctype_not_found
        - design_not_found
        - design_type_not_found
        - team_not_found
        - comment_not_found
        - too_many_comments
        - too_many_replies
        - message_too_long
        - design_not_fillable
        - autofill_data_invalid
        - feature_not_available
        - license_required
      type: string
...
    ExchangeAccessTokenRequest:
      discriminator:
        mapping:
          authorization_code: '#/components/schemas/ExchangeAuthCodeRequest'
          refresh_token: '#/components/schemas/ExchangeRefreshTokenRequest'
        propertyName: grant_type
      oneOf:
        - $ref: '#/components/schemas/ExchangeAuthCodeRequest'
        - $ref: '#/components/schemas/ExchangeRefreshTokenRequest'
      type: object
    ExchangeAuthCodeRequest:
      properties:
        grant_type:
          description: For exchanging an authorization code for an access token.
          enum:
            - authorization_code
          example: authorization_code
          type: string
        code_verifier:
          description: The `code_verifier` value that you generated when creating
            the user authorization URL.
          example: i541qdcfkb4htnork0w92lnu43en99ls5a48ittv6udqgiflqon8vusojojakbq4
          type: string
        code:
          description: The authorization code you received after the user authorized
            the integration.
          example: kp8nnroja7qnx00.opyc1p76rcbyflsxbycjqfp3ub8vzsvltpzwafy9q5l45dn5fxzhe7i7a6mg1i2t8jpsa6sebdeumkzzhicskabgevrxsssec4dvjwfvhq4gs3ugghguar0voiqpfb7axsapiojoter8v3w2s5s3st84jpv2l06h667iw241xngy9c8=vu1tnjp7sz
          type: string
        client_id:
          description: |
            Your integration's unique ID, for authenticating the request.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: OC-FAB12-AbCdEf
          type: string
        client_secret:
          description: |
            Your integration's client secret, for authenticating the request. Begins with `cnvca`.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: cnvcaAbcdefg12345_hijklm6789
          type: string
        redirect_uri:
          description: |
            Only required if a redirect URL was supplied when you [created the user authorization URL](https://www.canva.dev/docs/connect/authentication/#create-the-authorization-url).

            Must be one of those already specified by the client. If not supplied, the first redirect_uri defined for the client will be used by default.
          example: https://example.com/process-auth
          type: string
      required:
        - code
        - code_verifier
        - grant_type
      type: object
    ExchangeRefreshTokenRequest:
      properties:
        grant_type:
          description: For generating an access token using a refresh token.
          enum:
            - refresh_token
          example: refresh_token
          type: string
        client_id:
          description: |
            Your integration's unique ID, for authenticating the request.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: OC-FAB12-AbCdEf
          type: string
        client_secret:
          description: |
            Your integration's client secret, for authenticating the request. Begins with `cnvca`.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: cnvcaAbcdefg12345_hijklm6789
          type: string
        refresh_token:
          description: The refresh token to be exchanged. You can copy this value
            from the successful response received when generating an access token.
          example: JABix5nolsk9k8n2r0f8nq1gw4zjo40ht6sb4i573wgdzmkwdmiy6muh897hp0bxyab276wtgqkvtob2mg9aidt5d6rcltcbcgs101
          type: string
        scope:
          description: |
            Optional scope value when refreshing an access token. Separate multiple [scopes](https://www.canva.dev/docs/connect/appendix/scopes/) with a single space between each scope.

            The requested scope cannot include any permissions not already granted, so this parameter allows you to limit the scope when refreshing a token. If omitted, the scope for the token remains unchanged.
          example: design:meta:read
          type: string
      required:
        - grant_type
        - refresh_token
      type: object
    ExchangeAccessTokenResponse:
      description: Exchange auth token to access token.
      properties:
        access_token:
          description: "The bearer access token to use to authenticate to Canva Connect\
            \ API endpoints. If requested using a `authorization_code` or `refresh_token`,\
            \ this allows you to act on behalf of a user."
          example: JagALLazU0i2ld9WW4zTO4kaG0lkvP8Y5sSO206ZwxNF4E1y3xKJKF7TzN17BXTfaNOeY0P88AeRCE6cRF7SJzvf3Sx97rA80sGHtFplFo
          type: string
        refresh_token:
          description: The token that you can use to refresh the access token.
          example: JABix5nolsk9k8n2r0f8nq1gw4zjo40ht6sb4i573wgdzmkwdmiy6muh897hp0bxyab276wtgqkvtob2mg9aidt5d6rcltcbcgs101
          type: string
        token_type:
          description: The token type returned. This is always `Bearer`.
          example: Bearer
          type: string
        expires_in:
          description: The expiry time (in seconds) for the token.
          example: 14400
          format: int64
          type: integer
        scope:
          description: "The [scopes](https://www.canva.dev/docs/connect/appendix/scopes/)\
            \ that the token has been granted."
          example: asset:read design:meta:read design:permission:read folder:read
          type: string
      required:
        - access_token
        - expires_in
        - refresh_token
        - token_type
      type: object
...
    RevokeTokensRequest:
      description: Supply an access token or refresh token to have its lineage revoked.
      properties:
        token:
          description: The token to revoke.
          example: agALLazU0i2ld9WW4zTO4kaG0lkvP8Y5sSO206ZwxNF4E1y3xKJKF7TzN17BXTfaNOeY0P88AeRCE6cRF7SJzvf3Sx97rA80sGHtFplFo
          type: string
        client_id:
          description: |
            Your integration's unique ID, for authenticating the request.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: OC-FAB12-AbCdEf
          type: string
        client_secret:
          description: |
            Your integration's client secret, for authenticating the request. Begins with `cnvca`.

            NOTE: We recommend that you use basic access authentication instead of specifying `client_id` and `client_secret` as body parameters.
          example: cnvcaAbcdefg12345_hijklm6789
          type: string
      required:
        - token
      type: object
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http
    oauthAuthCode:
      flows:
        authorizationCode:
          authorizationUrl: https://www.canva.com/oauth/api/authorize
          scopes:
            design:content:read: View the contents of the user's designs.
            design:meta:read: View the metadata of the user's designs.
            design:content:write: Create designs on the user's behalf.
            design:permission:read: View the permissions assigned to a user's designs.
            design:permission:write: "Set, update, or remove the permissions assigned\
              \ to the user's designs."
            folder:read: |-
              View the metadata and contents of the user's folders, including their **Projects**
              folder.
            folder:write: |-
              Add, move, or remove the user's folders. It also lets you edit folder metadata,
              such as the folder's name.
            folder:permission:read: View permissions assigned to a user's folders.
            folder:permission:write: "Set, update, or remove permissions assigned\
              \ to the user's folders."
            asset:read: "View the metadata for the user's assets, such as uploaded\
              \ images."
            asset:write: "Upload, update, or delete assets on the user's behalf."
            comment:read: "View the comments on the user's designs, and the associated\
              \ metadata."
            comment:write: Create comments and replies on the user's designs.
            collaboration:event: Receive webhook notifications about events relevant
              to the user.
            brandtemplate:meta:read: View the metadata of the brand templates associated
              with the user's brand.
            brandtemplate:content:read: Read the content of the brand templates associated
              with the user's brand.
            profile:read: Read a user's profile and account information.
          tokenUrl: https://api.canva.com/rest/v1/oauth/token
      type: oauth2